Rewrite of ChoiceScript Monarch tokenizer
Complete rewrite of the majority of the
choicescript.ts file. It now uses two primary
states: one for plain text and one for command
lines. It switches between these by detecting
the patterns at the start of the next line.
It'd be tidier to switch back to a root state
via EOL, but Monarch doesn't seem to be able to
support that very well.
See the following Github issue:
https://github.com/microsoft/monaco-editor/issues/2265

Otherwise this rewrite greatly enhances the
tokenzior; providing support for much more
targeted tokenizing (and thus highlighting),
particularly in regard to strings, plain text
lines and expressions.